bpf: Open map on recovery if needed
If the first BPF map operation performed on a map would open the map and
that fails, the queued error would not open the map on recovery.
Recovery would be postponed until a subsequent error triggers another
open.

Fix this by opening the map if needed in the recovery synchronization.

Signed-off-by: Thomas Graf <thomas@cilium.io>