hubble/parser: Populate traffic direction for TraceTo{Lxc,Host,Stack}
Where possible, populate the TrafficDirection field from the trace
notification. This is only possible for observation points for which we
a) statically know the traffic direction and b) have connection
tracking information available to filter out reply packages.

The above observation points seem to be the only three for which we can
do that.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>