bpf: Remove FROM_HOST macro
For the host firewall, we need to have a single bpf_netdev object file
containing both the from-host and the from-netdev paths, to load it from
the Go side. We therefore turn FROM_HOST into a variable set to true or
false depending on the entry point (i.e., from_host() vs. from_netdev()).

The other option (to avoid the increased code complexity in datapath) is
to let the Go side handle several object files and header files for the
special host endpoint. That second option is likely to just move the
increase code complexity to the Go side.

Signed-off-by: Paul Chaignon <paul@cilium.io>